hv: shell & vuart: Change interrupt pin to uint32_t
Change the type from uint8_t to uint32_t for shell and vuart

Tracked-On: #861
Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/release/vuart.c
@@ -16,7 +16,7 @@ struct acrn_vuart *vuart_console_active(void)
 void vuart_console_tx_chars(__unused struct acrn_vuart *vu) {}
 void vuart_console_rx_chars(__unused struct acrn_vuart *vu) {}
 
-bool hv_used_dbg_intx(__unused uint8_t intx_pin)
+bool hv_used_dbg_intx(__unused uint32_t intx_pin)
 {
 	return false;
 }
